Title: 최소 에너지 비용, Time: -1ms, Memory: -1MB, Status: Compile Error - Codetree

+++ 250403/최소 에너지 비용/minimum-energy-cost.cpp
@@ -15,7 +15,7 @@ int main() {
     }
     long long sum = 0;
     sum = arr[0] * dis[0];
-    int min_cost = arr[0];
+    long long min_cost = arr[0];
     for(int i = 1; i<n-1;i++){
         min_cost = min(min_cost,arr[i]);
         sum += min_cost * dis[i];
